fixed all issues so results are personal

+++ music-App/backend/models/journalEntry.js
@@ -9,18 +9,18 @@ export async function addJournalEntry(journalEntry) {
   const { insertedId } = await collection.insertOne(journalEntry)
   return insertedId
 }
-export async function getJournalEntries() {
+export async function getJournalEntries(userId) {
   const collection = await getCollection()
-  const entries = await collection.find().toArray()
+  const entries = await collection.find({ userId }).toArray()
   return entries.map(entry => ({
     ...entry,
     _id: entry._id.toString() 
   }))
 }
 
-export async function deleteJournalEntry(id) {
+export async function deleteJournalEntry(id,userId) {
   const collection = await getCollection()
   const _id = new ObjectId(id) // Convert string ID to ObjectId
-  const result = await collection.deleteOne({ _id })
+  const result = await collection.deleteOne({ _id , userId })
   return result.deletedCount > 0 
 }
